backend-py/requirements.txt
fastapi
uvicorn
sqlalchemy
crewai
openai
python-dotenv
pymongo


# pip install -r requirements.txt